Create utils.py
